+++ pennylane_lightning/core/utils/cuda_utils/cuGates_host.hpp
@@ -1727,8 +1727,6 @@ template <class PrecisionT> class DynamicGateDataAccess {
     using NonParamGateFuncMap =
         std::unordered_map<std::string, NonParamGateFunc>;
 
-    // TODO: Need changes to support to the controlled gate tensor API once the
-    // API is finalized in cutensornet lib.
     NonParamGateFuncMap nonparametric_gates_{
         {"Identity",
          []() -> std::vector<CFP_t> { return cuGates::getIdentity<CFP_t>(); }},
@@ -1792,8 +1790,6 @@ template <class PrecisionT> class DynamicGateDataAccess {
         {"CSWAP",
          []() -> std::vector<CFP_t> { return cuGates::getCSWAP<CFP_t>(); }}};
 
-    // TODO: Need changes to support to the controlled gate tensor API once the
-    // API is finalized in cutensornet lib.
     ParamGateFuncMap parametric_gates_{
         {"PhaseShift",
          [](auto &&params) {