+++ scripts/configure_pyproject_toml.py
@@ -104,9 +104,6 @@ if __name__ == "__main__":
     if backend == "lightning_gpu":
         cuda_version = cuda_version()
         requires.append(f"custatevec-cu{cuda_version}")
-    if backend == "lightning_tensor":
-        cuda_version = cuda_version()
-        requires.append(f"cutensornet-cu{cuda_version}")
 
     pyproject["build-system"]["requires"] = requires
 
@@ -136,10 +133,6 @@ if __name__ == "__main__":
     if backend == "lightning_gpu":
         dependencies += [f"custatevec-cu{cuda_version}"]
 
-    if backend == "lightning_tensor":
-        dependencies += [f"cutensornet-cu{cuda_version}", f"nvidia-cusolver-cu{cuda_version}"]
-
-    if backend in ("lightning_gpu", "lightning_tensor"):
         # We only need to append "12" for CUDA 12 packages as these
         # specific Nvidia libraries currently do not append "13" for CUDA 13
         package_suffix = "-cu12" if cuda_version == "12" else ""