+++ tests/lightning_base/test_simulate_method.py
@@ -22,9 +22,6 @@ import pytest
 from conftest import PHI, THETA, LightningDevice, device_name  # tested device
 from pennylane.devices import DefaultQubit, ExecutionConfig
 
-if device_name == "lightning.tensor":
-    pytest.skip("Skipping tests for the LightningTensor class.", allow_module_level=True)
-
 if not LightningDevice._CPP_BINARY_AVAILABLE:
     pytest.skip("No binary module found. Skipping.", allow_module_level=True)