+++ tests/test_vjp.py
@@ -28,9 +28,6 @@ from pennylane.exceptions import QuantumFunctionError
 if not ld._CPP_BINARY_AVAILABLE:
     pytest.skip("No binary module found. Skipping.", allow_module_level=True)
 
-if device_name == "lightning.tensor":
-    pytest.skip("lightning.tensor doesn't support vjp.", allow_module_level=True)
-
 
 class TestVectorJacobianProduct:
     """Tests for the `vjp` function"""